Move update notifications to the CLI folder
It feels natural to make notifications an actual concern of the CLI. This is especially true as I am going to be adding birthday notifications too.
