+++ src/cli/notifications/shouldShowNotifications.ts
@@ -1,9 +1,9 @@
 import { parseBoolean } from "@alextheman/utility";
 
-const shouldRunAutomatedUpdateCheck = !(
+const shouldShowNotifications = !(
   process.env.NODE_ENV === "test" ||
   parseBoolean(process.env.RUN_END_TO_END ?? "false") ||
   parseBoolean(process.env.CI ?? "false")
 );
 
-export default shouldRunAutomatedUpdateCheck;
+export default shouldShowNotifications;
