+++ src/cli/notifications/updates/runAutomatedUpdateCheck.ts
@@ -2,8 +2,8 @@ import { ONE_DAY_IN_MILLISECONDS } from "@alextheman/utility";
 
 import createAlexCLineGlobalCache from "src/cache/global/createAlexCLineGlobalCache";
 import loadAlexCLineGlobalCache from "src/cache/global/loadAlexCLineGlobalCache";
-import checkUpdate from "src/utility/updates/checkUpdate";
-import { setPendingUpdateMessage } from "src/utility/updates/pendingUpdateMessage";
+import checkUpdate from "src/cli/notifications/updates/checkUpdate";
+import { setPendingUpdateMessage } from "src/cli/notifications/updates/pendingUpdateMessage";
 
 import { version } from "package.json" with { type: "json" };
